% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_Theta.R
\name{list_Thetas}
\alias{list_Thetas}
\title{List Unique Theta Names}
\usage{
list_Thetas(PMLParametersSets, IncludeAll = FALSE, IncludeCustom = TRUE)
}
\arguments{
\item{PMLParametersSets}{\code{PMLModels} class instance or an element (one PML
structure) of this class or \code{StParm} class.}

\item{IncludeAll}{Logical. Whether should the Theta names to be inlcuded from
structural parameters, covariates or thetas with a \code{State == 'None'}.}

\item{IncludeCustom}{Logical. Should the names of custom \code{theta} statements
(from the PML code of custom spaces) be included or not. Default is \code{TRUE}.}
}
\value{
A character vector containing the unique names of Theta parameters.
}
\description{
This function lists the unique names of Theta parameters in a given set.
}
\examples{
PMLParametersSets <- create_ModelPD()
list_Thetas(PMLParametersSets)

}
\seealso{
\code{\link[=Theta]{Theta()}} \code{\link[=modify_Theta]{modify_Theta()}}
}
