% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLStructureCustom.R
\name{output.CustomSpace}
\alias{output.CustomSpace}
\title{Output a Custom Space}
\usage{
\method{output}{CustomSpace}(x, ...)
}
\arguments{
\item{x}{A \code{CustomSpace} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string containing the PML code.
}
\description{
This function generates the PML code representation of a custom space.
}
