% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyDarwinOptionsPostprocess.R
\name{pyDarwinOptionsPostprocess}
\alias{pyDarwinOptionsPostprocess}
\title{Create pyDarwin Postprocess Options}
\usage{
pyDarwinOptionsPostprocess(
  use_r = FALSE,
  post_run_r_code = "{project_dir}/simplefunc.R",
  r_timeout = 30,
  use_python = FALSE,
  post_run_python_code = "{project_dir}/simplefunc.py"
)
}
\arguments{
\item{use_r}{Logical: Whether to use R for postprocessing. If set to TRUE, R
will be used to execute the post-processing script specified in
\code{post_run_r_code}. Default: FALSE.}

\item{post_run_r_code}{Character: The file path to the R script that contains
post-processing code. This script will be executed after the pyDarwin
optimization process finishes. It can perform additional analysis or
manipulations on the generated results.}

\item{r_timeout}{Numeric: The time limit (in seconds) for the execution of
the post-processing R script. If the R script takes longer to execute than
this timeout value, it will be terminated. Default: 30}

\item{use_python}{Logical: Whether to use Python for postprocessing. If set
to TRUE, Python will be used to execute the post-processing script
specified in \code{post_run_python_code}. Default: FALSE}

\item{post_run_python_code}{Character: The file path to the Python script
that contains post-processing code. This script will be executed after the
pyDarwin optimization process finishes. It can perform additional analysis
or manipulations on the generated results. Default:
\{project_dir\}/simplefunc.py}
}
\value{
A list of postprocessing options in pyDarwin optimization process.
}
\description{
Generates a list of postprocessing options to be used in pyDarwin
optimization process.
}
\examples{
# Create postprocess options with default values
postprocess_options <- pyDarwinOptionsPostprocess()
# Create postprocess options with custom values
postprocess_options_custom <-
  pyDarwinOptionsPostprocess(use_r = TRUE,
                             post_run_r_code = "{project_dir}/postprocess.R",
                             r_timeout = 60,
                             use_python = TRUE,
                             post_run_python_code = "{project_dir}/postprocess.py")

}
