% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pyDarwin.R
\name{stop_pyDarwin}
\alias{stop_pyDarwin}
\title{Stop pyDarwin Model Search}
\usage{
stop_pyDarwin(
  InterpreterPath,
  Flags = c("-u", "-m"),
  ForceStop = FALSE,
  DirectoryPath = "."
)
}
\arguments{
\item{InterpreterPath}{Path to the Python interpreter executable.}

\item{Flags}{Flags to pass to the Python interpreter. Refer to Python
documentation for details. Note that \code{-m} is essential (runs library module
as a script and terminates option list).}

\item{ForceStop}{Logical. If \code{TRUE}, \code{-f} flag is added to force stop search
immediately.- Default is \code{FALSE}.}

\item{DirectoryPath}{the \code{DirectoryPath} argument of \code{\link[=run_pyDarwin]{run_pyDarwin()}} or the
parent folder of options file passed to that function.
Default is current working directory.}
}
\value{
Returned code of \code{\link[=system2]{system2()}} call.
}
\description{
This function stops a pyDarwin model search.
}
\examples{
\dontrun{
stop_pyDarwin(
  InterpreterPath = "~/darwin/venv/bin/python",
  DirectoryPath = "~/project_folder")
}

}
