% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLStructureCustom.R
\name{add_CustomSpace}
\alias{add_CustomSpace}
\title{Add a Custom Space}
\usage{
add_CustomSpace(Spaces, CustomCode)
}
\arguments{
\item{Spaces}{A list of existing spaces.}

\item{CustomCode}{A character string containing the custom code for the new space.}
}
\value{
A list of spaces with the new custom space added.
}
\description{
This function adds a custom space to a list of spaces.
}
